pyproject.toml
[project]
name = "renex"
version = "0.1.0"
description = "Backend Services for RenEx, A energy Swapping application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.3",
    "fastapi>=0.121.3",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
    "structlog>=25.5.0",
    "taskipy>=1.14.1",
    "tzdata>=2025.2",
    "uvicorn>=0.38.0",
]

[tool.taskipy.tasks]
api = { cmd = "uvicorn main:app --host=0.0.0.0 --port=8080"}
